move most of the dl offer code to DataLayer (and dedupe)
